feat: Refactor card stack layout logic and add flip animations
